Begin rewriting moderation commands
