+++ src/main/kotlin/org/hyacinthbots/lilybot/extensions/moderation/ModerationActions.kt
@@ -1,0 +1,9 @@
+package org.hyacinthbots.lilybot.extensions.moderation
+
+enum class ModerationActions {
+	BAN,
+	SOFT_BAN,
+	KICK,
+	WARN,
+	TIMEOUT
+}
